requirements.txt
numpy==1.23.0
#opencv_python==4.8.0.74
#opencv_python_headless==4.8.0.74
Pillow==10.0.0
#torch==2.0.1
#torchvision==0.15.2
tqdm==4.65.0
ultralytics==8.0.148
scikit-learn==1.3.0

# Server requirements:
fastapi
uvicorn
python-multipart